MDL-22787 Fixed enrol_mnet upgrade procedure
For any courses where at least one remote student was enrolled via 1.x
enrol_mnet plugin, the record in the table enrol was created
automatically by core during the upgrade to the new enrolment framework.
We just need to set the custom field to explicitly indicate that the
course is open for all remote hosts.